bump substrate to polkadotv0.9.30 (#273)

+++ pallets/hasher/src/weights.rs
@@ -55,8 +55,8 @@ impl<T: frame_system::Config> WeightInfo for WebbWeight<T> {
 	// Storage: HasherBn254 Parameters (r:1 w:1)
 	fn force_set_parameters(_c: u32, ) -> Weight {
 		Weight::from_ref_time(8_102_000)
-			.saturating_add(T::DbWeight::get().reads(1 as u64))
-			.saturating_add(T::DbWeight::get().writes(1 as u64))
+			.saturating_add(T::DbWeight::get().reads(1_u64))
+			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
 }
 
@@ -65,7 +65,7 @@ impl WeightInfo for () {
 	// Storage: HasherBn254 Parameters (r:1 w:1)
 	fn force_set_parameters(_c: u32, ) -> Weight {
 		Weight::from_ref_time(8_102_000)
-			.saturating_add(RocksDbWeight::get().reads(1 as u64))
-			.saturating_add(RocksDbWeight::get().writes(1 as u64))
+			.saturating_add(RocksDbWeight::get().reads(1_u64))
+			.saturating_add(RocksDbWeight::get().writes(1_u64))
 	}
 }
