+++ pallets/vanchor-verifier/src/weights.rs
@@ -57,8 +57,8 @@ impl<T: frame_system::Config> WeightInfo for WebbWeight<T> {
 		Weight::from_ref_time(3_653_000)
 			// Standard Error: 0
 			.saturating_add(Weight::from_ref_time(1_000).saturating_mul(c as u64))
-			.saturating_add(T::DbWeight::get().reads(1 as u64))
-			.saturating_add(T::DbWeight::get().writes(1 as u64))
+			.saturating_add(T::DbWeight::get().reads(1_u64))
+			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
 }
 
@@ -69,7 +69,7 @@ impl WeightInfo for () {
 		Weight::from_ref_time(3_653_000)
 			// Standard Error: 0
 			.saturating_add(Weight::from_ref_time(1_000).saturating_mul(c as u64))
-			.saturating_add(RocksDbWeight::get().reads(1 as u64))
-			.saturating_add(RocksDbWeight::get().writes(1 as u64))
+			.saturating_add(RocksDbWeight::get().reads(1_u64))
+			.saturating_add(RocksDbWeight::get().writes(1_u64))
 	}
 }
